[RHPAM-3681] - Introduce BuildContext
Create separate handler for buildconfig

Update release notes

Fix comment

Implement decorator for custom labels
