+++ RELEASE_NOTES.md
@@ -1,15 +1,7 @@
 <!-- Keep them in alphabetical order -->
 ## Enhancements
-- [KOGITO-4455](https://issues.redhat.com/browse/KOGITO-4455) - Make Kogito Images/CLI to build a Kogito Workflow project from a workflow definition file
-- [KOGITO-4722](https://issues.redhat.com/browse/KOGITO-4722) - Support Job Service MongoDB distribution provisioning in Operator
-- [KOGITO-5315](https://issues.redhat.com/browse/KOGITO-5315) - Jobs service use unsupported authentication mechanism (part of Infinispan 12 upgrade - https://issues.redhat.com/browse/KOGITO-4341)
+- [RHPAM-3681](https://issues.redhat.com/browse/RHPAM-3681) - Missing metering labels for Kogito pods
 
 ## Bug Fixes
-- [KOGITO-4453](https://issues.redhat.com/browse/KOGITO-4453) - Kogito operator overwrites user's configmap provided with propertiesConfigMap
-- [KOGITO-4898](https://issues.redhat.com/browse/KOGITO-4898) - Probe defaults should be based on YAML action values
-- [KOGITO-5005](https://issues.redhat.com/browse/KOGITO-5005) - Operator print irrelevant metadata in logs
-- [KOGITO-5104](https://issues.redhat.com/browse/KOGITO-5104) - Kogito operator creates wrong default image tags
-- [KOGITO-5083](https://issues.redhat.com/browse/KOGITO-5083) - Kogito Runtime Image Stream Fails to Update
-- [KOGITO-5256](https://issues.redhat.com/browse/KOGITO-5256) - Kogito Operator is creating incorrect Knative Triggers
 
 ## Known Issues
