+++ core/kogitobuild/status.go
@@ -19,7 +19,6 @@ import (
 	"github.com/kiegroup/kogito-operator/core/client/kubernetes"
 	"github.com/kiegroup/kogito-operator/core/client/openshift"
 	"github.com/kiegroup/kogito-operator/core/framework"
-	"github.com/kiegroup/kogito-operator/core/operator"
 	buildv1 "github.com/openshift/api/build/v1"
 	"k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -45,13 +44,13 @@ type StatusHandler interface {
 }
 
 type statusHandler struct {
-	operator.Context
+	BuildContext
 }
 
 // NewStatusHandler ...
-func NewStatusHandler(context operator.Context) StatusHandler {
+func NewStatusHandler(context BuildContext) StatusHandler {
 	return &statusHandler{
-		Context: context,
+		context,
 	}
 }
