+++ core/kogitobuild/status_test.go
@@ -45,10 +45,12 @@ func TestStatusChangeWhenConsecutiveErrorsOccur(t *testing.T) {
 	}
 	cli := test.NewFakeClientBuilder().AddK8sObjects(instance).Build()
 	err := errors.New("error")
-	context := operator.Context{
-		Client: cli,
-		Log:    test.TestLogger,
-		Scheme: meta.GetRegisteredSchema(),
+	context := BuildContext{
+		Context: operator.Context{
+			Client: cli,
+			Log:    test.TestLogger,
+			Scheme: meta.GetRegisteredSchema(),
+		},
 	}
 	buildStatusHandler := NewStatusHandler(context)
 	buildStatusHandler.HandleStatusChange(instance, err)
@@ -85,12 +87,14 @@ func TestStatusChangeWhenBuildsAreRunning(t *testing.T) {
 		},
 	}
 	cli := test.NewFakeClientBuilder().OnOpenShift().AddK8sObjects(instance).Build()
-	context := operator.Context{
-		Client: cli,
-		Log:    test.TestLogger,
-		Scheme: meta.GetRegisteredSchema(),
+	context := BuildContext{
+		Context: operator.Context{
+			Client: cli,
+			Log:    test.TestLogger,
+			Scheme: meta.GetRegisteredSchema(),
+		},
 	}
-	deltaProcessor := &deltaProcessor{Context: context, build: instance}
+	deltaProcessor := &deltaProcessor{BuildContext: context, build: instance}
 	manager := deltaProcessor.getBuildManager()
 	requested, err := manager.GetRequestedResources()
 	assert.NoError(t, err)
@@ -153,10 +157,12 @@ func TestStatusChangeWhenBuildsAreRunning(t *testing.T) {
 	// recreating the Client with our objects to make sure that the BCs will be there
 	cli = test.NewFakeClientBuilder().AddK8sObjects(k8sObjs...).AddBuildObjects(buildObjs...).Build()
 	err = nil
-	context1 := operator.Context{
-		Client: cli,
-		Log:    test.TestLogger,
-		Scheme: meta.GetRegisteredSchema(),
+	context1 := BuildContext{
+		Context: operator.Context{
+			Client: cli,
+			Log:    test.TestLogger,
+			Scheme: meta.GetRegisteredSchema(),
+		},
 	}
 	buildStatusHandler := NewStatusHandler(context1)
 	buildStatusHandler.HandleStatusChange(instance, err)
