+++ internal/kogitobuild.go
@@ -18,19 +18,19 @@ import (
 	"github.com/kiegroup/kogito-operator/api"
 	"github.com/kiegroup/kogito-operator/api/v1beta1"
 	"github.com/kiegroup/kogito-operator/core/client/kubernetes"
+	"github.com/kiegroup/kogito-operator/core/kogitobuild"
 	"github.com/kiegroup/kogito-operator/core/manager"
-	"github.com/kiegroup/kogito-operator/core/operator"
 	"k8s.io/apimachinery/pkg/types"
 )
 
 type kogitoBuildHandler struct {
-	operator.Context
+	kogitobuild.BuildContext
 }
 
 // NewKogitoBuildHandler ...
-func NewKogitoBuildHandler(context operator.Context) manager.KogitoBuildHandler {
+func NewKogitoBuildHandler(context kogitobuild.BuildContext) manager.KogitoBuildHandler {
 	return &kogitoBuildHandler{
-		Context: context,
+		context,
 	}
 }
